# perf(runtime): optimize Deno.file open & stream