Opening and closing files is 2.5x faster
```
# Deno new
open_file_sync     7.55 µs/iter     (7.51 µs … 7.72 µs)   7.56 µs   7.72 µs   7.72 µs
# Deno `main`
open_file_sync    18.31 µs/iter   (17.79 µs … 39.75 µs)  18.25 µs  19.62 µs  20.62 µs
```

ext/flash send_file benchmark now gives 7GB/s transfer rate. 1.3x faster than `main`:
```
Transfer/sec:      5.79GB # main
Transfer/sec:      7.56GB # new
```