@aapoalas 2022-08-19
> Opening and closing files is 2.5x faster
> 
> ```
> # Deno `main`
> open_file_sync     7.55 µs/iter     (7.51 µs … 7.72 µs)   7.56 µs   7.72 µs   7.72 µs
> # Deno new
> open_file_sync    18.31 µs/iter   (17.79 µs … 39.75 µs)  18.25 µs  19.62 µs  20.62 µs
> ```

Am I reading this wrong somehow? I read this as the new `open_file_sync` taking 18 microseconds and the old one taking only 7.5 microseconds, ie. old one being faster.