@littledivy 2022-08-19
> Am I reading this wrong somehow? I read this as the new open_file_sync taking 18 microseconds and the old one taking only 7.5 microseconds, ie. old one being faster.

Oopsie, wrong labels.